Fix the class design
Fixing the static class to be OOP design
