+++ src/main/java/org/coffeemine/app/spring/CalendarComponent.java
@@ -1,74 +1,47 @@
 package org.coffeemine.app.spring;
 
-import com.vaadin.flow.component.Text;
-import com.vaadin.flow.component.notification.Notification;
-import com.vaadin.flow.component.orderedlayout.VerticalLayout;
-import org.vaadin.stefan.fullcalendar.CalendarViewImpl;
-import org.vaadin.stefan.fullcalendar.Entry;
-import org.vaadin.stefan.fullcalendar.FullCalendar;
+import com.vaadin.flow.component.button.Button;
+import com.vaadin.flow.component.datepicker.DatePicker;
+import com.vaadin.flow.component.formlayout.FormLayout;
+import com.vaadin.flow.component.orderedlayout.HorizontalLayout;
+import com.vaadin.flow.component.select.Select;
+import com.vaadin.flow.component.textfield.TextField;
 
-import java.time.DayOfWeek;
 import java.time.LocalDate;
-import java.time.LocalDateTime;
 
-public class CalendarComponent extends VerticalLayout {
+class CalendarComponent extends HorizontalLayout {
 
-    private static FullCalendar systemCalendar;
-    private Text date;
+    CalendarComponent(CalendarForm calendarForm){
+        Button today = new Button("Today", event -> calendarForm.calendarGoToday());
+        Button previous = new Button("<< Previous", event -> calendarForm.calendarGoPrevious());
+        Button next = new Button("next >>",event -> calendarForm.calendarGoNext());
+        Select<String> selectView = new Select<>("Month", "Week", "Day");
+        selectView.setPlaceholder("View as..");
+        selectView.addValueChangeListener(event -> calendarForm.changeCalendarView(event.getValue()));
+        Button newTask = new Button("New task", event -> calendarForm.createNewTask());
+        DatePicker datePicker = new DatePicker();
 
-    public CalendarComponent(){
-        systemCalendar = new FullCalendar();
-        date = new Text(LocalDateTime.now().getYear() + "  " + LocalDateTime.now().getMonth().toString());
-        systemCalendar.setFirstDay(DayOfWeek.MONDAY);
-        systemCalendar.setBusinessHours();
-        add(date);
-        add(systemCalendar);
+        datePicker.setPlaceholder("Date within this month");
+        LocalDate nowDate = LocalDate.now();
+
+        datePicker.setClearButtonVisible(true);
+
+        datePicker.addValueChangeListener( event -> {
+            LocalDate newDate = event.getValue();
+            calendarForm.calendarGoToDate(newDate);
+        });
+
+        add(previous);
+        add(today);
+        add(next);
+        add(selectView);
+        add(newTask);
+        add(datePicker);
     }
 
-    static void calendarGoNext(){
-        systemCalendar.next();
-    }
-
-    static void calendarGoPrevious(){
-        systemCalendar.previous();
-    }
-
-    static void calendarGoToday(){
-        systemCalendar.today();
-    }
-
-    static void changeCalendarView(String viewName){
-        switch(viewName) {
-            case "Day":
-                systemCalendar.changeView(CalendarViewImpl.DAY_GRID_DAY);
-                Notification.show(viewName + " view now");
-                break;
-            case "Week":
-                systemCalendar.changeView(CalendarViewImpl.DAY_GRID_WEEK);
-                Notification.show(viewName + " view now");
-                break;
-            default:
-                systemCalendar.changeView(CalendarViewImpl.DAY_GRID_MONTH);
-                Notification.show(viewName + " view now");
-                break;
-        }
-    }
-
-    static void createNewTask(){
-        Entry newEvent = new Entry(
-                "0101",
-                "Nice Calendar View",
-                LocalDateTime.now(),
-                LocalDateTime.of(2019,11,28,12,00),
-                true,
-                true,
-                " Deep blue",
-                "This is a demo task");
-        systemCalendar.addEntries(newEvent);
-        Notification.show("Task " + "<Nice Calendar View>" + " is now added");
-    }
-
-    static void calendarGoToDate(LocalDate date) {
-        systemCalendar.gotoDate(date);
+    FormLayout newTaskForm() {
+        FormLayout newForm = new FormLayout();
+        newForm.addFormItem(new TextField(), "Task Name");
+        return newForm;
     }
 }
